requirements.txt
pynacl==1.5.0
pytest==7.4.0